# backend/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gram Sabha Management System API",
    "description": "\n## Comprehensive API for Digital Village Governance\n\nThis API serves both Admin Portal (panchayat management) and Citizen Portal (services & issue reporting).\n\n### Features:\n- **Biometric Authentication** - Facial recognition login\n- **Panchayat & Ward Management** - Complete CRUD operations\n- **Issue Tracking** - Rich content reporting with images/audio\n- **Multi-language Support** - English & Hindi\n- **Member Management** - CSV import, bulk operations\n\n### Authentication:\nMost endpoints require JWT Bearer token authentication.\n\n### File Uploads:\nEndpoints support multipart/form-data for images, audio, and documents.\n    ",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://egramsabha.empowerpanchayat.org/",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT Bearer token"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/jiosign/doc-status/{groupid}": {
      "get": {
        "tags": [
          "Jio Signs"
        ],
        "summary": "",
        "parameters": [
          {
            "name": "groupid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter: groupid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/jiosign/signed-file/{groupid}": {
      "get": {
        "tags": [
          "Jio Signs"
        ],
        "summary": "",
        "parameters": [
          {
            "name": "groupid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter: groupid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  }
}